Remove Robot#context in favor of Plugin.context
This change was made to avoid robot instances having to know anything
about plug-ins, specifically having to record the plug-in state in order
to associate capabilities with plug-ins. Unfortunately, now capabilities
have to know about plug-ins. This is somewhat preferable, since
capabilities already associate themselves with plug-in objects. Plug-in
state is now stored and queried on the Plugin class.